KS-14-2 move Classes in Main.kt
